Fixed Queued view and prepare native ads view

+++ app/src/main/java/com/zsk/androtweet2/activities/MainActivity.kt
@@ -86,7 +86,7 @@ open class MainActivity : BaseActivity(), Drawer.OnDrawerItemClickListener, Acco
                     androTweetApp.accountHeader.let { accountHeader ->
                         dataSnapShot?.getValue<TwitterAccount>(TwitterAccount::class.java)
                                 ?.let { account ->
-                                    accountHeader.removeProfileByIdentifier(account.id.toLong())
+                                    accountHeader.removeProfileByIdentifier(account.id)
                                 }
                     }
                 }
@@ -97,7 +97,7 @@ open class MainActivity : BaseActivity(), Drawer.OnDrawerItemClickListener, Acco
     }
 
     private fun getProfileDrawerItem(account: TwitterAccount): IProfile<*> {
-        return ProfileDrawerItem().withIdentifier(account.id.toLong())
+        return ProfileDrawerItem().withIdentifier(account.id)
                 .withName(account.name)
                 .withEmail(account.realname)
                 .withIcon(account.profilePic)
